Register only the endpoint 0x01
We used to register each endpoint using the same controlBridge structure.
However in the zcl.c, we override to endpoint 0x01. Thus registering would just increase heap footprint with no additional benefit
We will have to check in zcl_search if we want to maintain a list of endpoint in eZCL_SearchForEPIndex or just send back the index of endpoint 0x01 everytime